refactor route subjects server
