+++ server/test/academy.test.js
@@ -305,771 +305,6 @@ describe('#PUT /class', () => {
   });
 });
 
-describe('#POST /academy/subject', () => {
-  let connect;
-  let createSubjectStub;
-  let query;
-  let findOneStub;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    createSubjectStub = sinon.stub(network, 'createSubject');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    createSubjectStub.restore();
-    query.restore();
-  });
-
-  it('do not create success subject with empty req.body', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectName: '',
-        subjectCode: '',
-        shortDescription: '',
-        description: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not create success subject with empty subjectName', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectName: '',
-        subjectCode: 'bc001',
-        shortDescription: 'lorem',
-        description: 'lorem...'
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not create success subject with empty subjectCode', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectName: 'php',
-        subjectCode: '',
-        shortDescription: 'lorem',
-        description: 'lorem...'
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not create success subject with empty shortDescription', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectName: 'php',
-        subjectCode: '123',
-        shortDescription: '',
-        description: 'lorem...'
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not create success subject with empty description', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectName: 'php',
-        subjectCode: '123',
-        shortDescription: 'lorem',
-        description: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not create success subject with admin student', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectName: 'blockchain',
-        subjectCode: 'bc001',
-        shortDescription: 'lorem',
-        description: 'lorem...'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not create success subject with teacher', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .send({
-        subjectName: 'blockchain',
-        subjectCode: 'bc001',
-        shortDescription: 'lorem',
-        description: 'lorem...'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not create success subject with student', (done) => {
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .send({
-        subjectName: 'blockchain',
-        subjectCode: 'bc001',
-        shortDescription: 'lorem',
-        description: 'lorem...'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not create success subject with admin academy - error chaincode', (done) => {
-    createSubjectStub.returns({
-      success: false,
-      msg: 'error'
-    });
-
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subjectName: 'blockchain',
-        subjectCode: 'bc001',
-        shortDescription: 'lorem',
-        description: 'lorem...'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('error');
-        done();
-      });
-  });
-
-  it('create success subject with admin academy', (done) => {
-    let data = JSON.stringify(
-      {
-        SubjectID: '00',
-        SubjectCode: 'bc001',
-        SubjectName: 'Blockchain',
-        ShortDescription: 'lorem01',
-        Description: 'lorem01...'
-      },
-      {
-        SubjectID: '01',
-        SubjectCode: 'etc002',
-        SubjectName: 'ETC',
-        ShortDescription: 'lorem02',
-        Description: 'lorem02...'
-      }
-    );
-
-    createSubjectStub.returns({
-      success: true,
-      msg: 'created'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .post('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subjectName: 'blockchain02',
-        subjectCode: 'bc002',
-        shortDescription: 'lorem3',
-        description: 'lorem3...'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-});
-
-describe('#PUT /academy/subject', () => {
-  let connect;
-  let updateSubjectStub;
-  let query;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    updateSubjectStub = sinon.stub(network, 'updateSubjectInfo');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    updateSubjectStub.restore();
-    query.restore();
-  });
-
-  it('do not update success subject with empty req.body', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '',
-          subjectName: '',
-          subjectCode: '',
-          shortDescription: '',
-          description: ''
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not update success subject with empty subjectId', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '',
-          subjectName: 'php',
-          subjectCode: 'bc001',
-          shortDescription: 'lorem',
-          description: 'lorem...'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not update success subject with empty subjectName', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '00',
-          subjectName: '',
-          subjectCode: 'bc001',
-          shortDescription: 'lorem',
-          description: 'lorem...'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not update success subject with empty subjectCode', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '00',
-          subjectName: 'php',
-          subjectCode: '',
-          shortDescription: 'lorem',
-          description: 'lorem...'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not update success subject with empty shortDescription', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '00',
-          subjectName: 'php',
-          subjectCode: '123',
-          shortDescription: '',
-          description: 'lorem...'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not update success subject with empty description', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '00',
-          subjectName: 'php',
-          subjectCode: '123',
-          shortDescription: 'lorem',
-          description: ''
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('do not update success subject with admin student', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '00',
-          subjectName: 'blockchain',
-          subjectCode: 'bc001',
-          shortDescription: 'lorem',
-          description: 'lorem...'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not update success subject with teacher', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '00',
-          subjectName: 'blockchain',
-          subjectCode: 'bc001',
-          shortDescription: 'lorem',
-          description: 'lorem...'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not update success subject with student', (done) => {
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: '00',
-          subjectName: 'blockchain',
-          subjectCode: 'bc001',
-          shortDescription: 'lorem',
-          description: 'lorem...'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('No changes', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      SubjectID: 'a98f2f4e-6ef9-492b-96a6-c6f01364fecb',
-      SubjectName: 'abcc',
-      SubjectCode: 'abc',
-      ShortDescription: 'abc',
-      Description: 'abc'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    updateSubjectStub.returns({
-      success: false,
-      msg: 'error'
-    });
-
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: 'a98f2f4e-6ef9-492b-96a6-c6f01364fecb',
-          subjectName: 'abcc',
-          subjectCode: 'abc',
-          shortDescription: 'abc',
-          description: 'abc'
-        }
-      })
-      .then((res) => {
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('No changes!');
-        done();
-      });
-  });
-
-  it('error chain code', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let data = JSON.stringify({
-      SubjectID: 'a98f2f4e-6ef9-492b-96a6-c6f01364fecb',
-      SubjectName: 'abc',
-      SubjectCode: 'abc',
-      ShortDescription: 'abc',
-      Description: 'abc'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    updateSubjectStub.returns({
-      success: false,
-      msg: 'error'
-    });
-
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: 'a98f2f4e-6ef9-492b-96a6-c6f01364fecb',
-          subjectName: 'abcc',
-          subjectCode: 'abc',
-          shortDescription: 'abc',
-          description: 'abc'
-        }
-      })
-      .then((res) => {
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('error');
-        done();
-      });
-  });
-
-  it('update success subject with admin academy', (done) => {
-    let data = JSON.stringify({
-      SubjectID: 'a98f2f4e-6ef9-492b-96a6-c6f01364fecb',
-      SubjectCode: 'abc',
-      SubjectName: 'abc',
-      ShortDescription: 'abc',
-      Description: 'abc...'
-    });
-
-    updateSubjectStub.returns({
-      success: true,
-      msg: 'updated'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .put('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subject: {
-          subjectId: 'a98f2f4e-6ef9-492b-96a6-c6f01364fecb',
-          subjectName: 'abcc',
-          subjectCode: 'abc',
-          shortDescription: 'abc',
-          description: 'abc'
-        }
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-});
-
-describe('#DELETE /academy/subject', () => {
-  let connect;
-  let deleteSubjectStub;
-  let query;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    deleteSubjectStub = sinon.stub(network, 'deleteSubject');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    deleteSubjectStub.restore();
-    query.restore();
-  });
-
-  it('do not delete success subject with empty req.body', (done) => {
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({})
-      .then((res) => {
-        expect(res.status).equal(400);
-        done();
-      });
-  });
-
-  it('do not delete success subject with empty subjectId', (done) => {
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectId: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(400);
-        done();
-      });
-  });
-
-  it('do not delete success subject with admin student', (done) => {
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_STUDENT_EXAMPLE}`)
-      .send({
-        subjectId: '00'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not delete success subject with teacher', (done) => {
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .send({
-        subjectId: '00'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('do not delete success subject with student', (done) => {
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .send({
-        subjectId: '00'
-      })
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.msg).equal('Permission Denied');
-        done();
-      });
-  });
-
-  it('Failed connect to blockchain', (done) => {
-    connect.returns(null);
-
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.msg).equal('Failed connect to blockchain');
-        done();
-      });
-  });
-
-  it('Can not query chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    query.returns({ success: false });
-
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(404);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-
-  it('Can not delete this subject now!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let subject = JSON.stringify({
-      SubjectID: 'subject01',
-      Classes: ['class01']
-    });
-
-    query.returns({ success: true, msg: subject });
-
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(400);
-        expect(res.body.msg).equal('Can not delete this subject now!');
-        done();
-      });
-  });
-
-  it('Can not invoke chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let subject = JSON.stringify({
-      SubjectID: 'subject01',
-      Classes: null
-    });
-
-    query.returns({ success: true, msg: subject });
-
-    deleteSubjectStub.returns({ success: false });
-
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.msg).equal('Can not invoke chaincode!');
-        done();
-      });
-  });
-
-  it('Delete Successfully!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let subject = JSON.stringify({
-      SubjectID: 'subject01',
-      Classes: null
-    });
-
-    query.returns({ success: true, msg: subject });
-
-    deleteSubjectStub.returns({ success: true });
-
-    request(app)
-      .delete('/academy/subject')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        subjectId: 'subject01'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.msg).equal('Delete Successfully!');
-        done();
-      });
-  });
-});
-
 describe('#PUT /academy/startClass', () => {
   let connect;
   let query;
@@ -1293,208 +528,6 @@ describe('#PUT /academy/startClass', () => {
         expect(res.status).equal(200);
         expect(res.body.success).equal(true);
         expect(res.body.msg).equal('Start Successfully!');
-        done();
-      });
-  });
-});
-
-describe('#POST /academy/deleteClass', () => {
-  let connect;
-  let query;
-  let deleteClass;
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    deleteClass = sinon.stub(network, 'deleteClass');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    deleteClass.restore();
-    query.restore();
-  });
-
-  it('Permission denied when access routes with student!', (done) => {
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied!');
-        done();
-      });
-  });
-
-  it('Permission denied when access routes with teacher!', (done) => {
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_TEACHER_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(403);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Permission Denied!');
-        done();
-      });
-  });
-
-  it('Invalid Body!', (done) => {
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        classId: ''
-      })
-      .then((res) => {
-        expect(res.status).equal(422);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-
-  it('Failed connect to blockchain!', (done) => {
-    connect.returns(null);
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        classId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Failed connect to blockchain!');
-        done();
-      });
-  });
-
-  it('Can not query chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    query.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        classId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not query chaincode!');
-        done();
-      });
-  });
-
-  it('Can delete this class now!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let classInfo = JSON.stringify({
-      ClassID: '123456',
-      Status: 'Closed'
-    });
-
-    query.returns({
-      success: true,
-      msg: classInfo
-    });
-
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        classId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can delete this class now!');
-        done();
-      });
-  });
-
-  it('Can not invoke chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let classInfo = JSON.stringify({
-      ClassID: '123456',
-      Status: 'Open'
-    });
-
-    query.returns({
-      success: true,
-      msg: classInfo
-    });
-
-    deleteClass.returns({
-      success: false
-    });
-
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        classId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        expect(res.body.msg).equal('Can not invoke chaincode!');
-        done();
-      });
-  });
-
-  it('Can not invoke chaincode!', (done) => {
-    connect.returns({
-      contract: 'academy',
-      network: 'certificatechannel',
-      gateway: 'gateway',
-      user: { username: 'adminacademy', role: USER_ROLES.ADMIN_ACADEMY }
-    });
-
-    let classInfo = JSON.stringify({
-      ClassID: '123456',
-      Status: 'Open'
-    });
-
-    query.returns({
-      success: true,
-      msg: classInfo
-    });
-
-    deleteClass.returns({
-      success: true
-    });
-
-    request(app)
-      .post('/academy/deleteClass')
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .send({
-        classId: '123456'
-      })
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        expect(res.body.msg).equal('Delete Successfully!');
         done();
       });
   });
