+++ server/test/common.test.js
@@ -9,86 +9,6 @@ const USER_ROLES = require('../configs/constant').USER_ROLES;
 
 require('dotenv').config();
 
-describe('#GET /common/subject/:subjectId/classes', () => {
-  let connect;
-  let query;
-  let subjectId = '4ca7fc39-7523-424d-984e-87ea590cac68';
-
-  beforeEach(() => {
-    connect = sinon.stub(network, 'connectToNetwork');
-    query = sinon.stub(network, 'query');
-  });
-
-  afterEach(() => {
-    connect.restore();
-    query.restore();
-  });
-
-  it('success query classess of subject with admin', (done) => {
-    let data = JSON.stringify({
-      classCode: 'CACLC2',
-      room: 'Blockchain101',
-      time: '12122020',
-      shortDescription: 'short',
-      description: 'Blockchain'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .get(`/common/subject/${subjectId}/classes`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-
-  it('success query classess of subject with student', (done) => {
-    let data = JSON.stringify({
-      classCode: 'CACLC2',
-      room: 'Blockchain101',
-      time: '12122020',
-      shortDescription: 'short',
-      description: 'Blockchain'
-    });
-
-    query.returns({
-      success: true,
-      msg: data
-    });
-
-    request(app)
-      .get(`/common/subject/${subjectId}/classes`)
-      .set('authorization', `${process.env.JWT_STUDENT_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(200);
-        expect(res.body.success).equal(true);
-        done();
-      });
-  });
-
-  it('do not success query class because error query chaincode', (done) => {
-    query.returns({
-      success: false,
-      msg: 'Error query chaincode'
-    });
-
-    request(app)
-      .get(`/common/subject/${subjectId}/classes`)
-      .set('authorization', `${process.env.JWT_ADMIN_ACADEMY_EXAMPLE}`)
-      .then((res) => {
-        expect(res.status).equal(500);
-        expect(res.body.success).equal(false);
-        done();
-      });
-  });
-});
-
 describe('#GET /common/subject/:subjectId', () => {
   let connect;
   let query;
